+++ src/main/java/com/example/springLoan/factory/ProductSettingFactoryImpl.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class ProductSettingFactoryImpl implements ProductSettingFactory {
     @Override
-    public ProductSetting getProductSetting(Integer id, String value, Product product, Setting setting) {
+    public ProductSetting getProductSetting(Long id, String value, Product product, Setting setting) {
         return new ProductSetting(id, value, product, setting);
     }
 }
